pyproject.toml
[project]
name = "kafka-study"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.134.0",
    "sqlmodel>=0.0.37",
    "motor>=3.3.2",
    "pymongo>=4.6.0",
    "python-dotenv>=1.0.0",
]
